+++ kod/gui/screen.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+class Screen(Enum):
+    MAIN = auto()
+    TOPSIS = auto()
+    RSM = auto()
+    UTA = auto()
+    SP = auto()
